// lib/types/hometax/TaxAccountantRetrieveRegSession.ts
export interface TaxAccountantRetrieveRegSessionResponse {
    errCode: string;
    errMsg: string;
    result: string;
    data: TaxAccountantRetrieveRegSessionResponseData;
}

export interface TaxAccountantRetrieveRegSessionResponseData {
    RESULT: string;
    ERRMSG: string;
    ERRDOC: string;
    ETRACK: string;
    ECODE: string;
}